Implements Val.lazy(), proposes Val.lazyAsync()
An implementation and documentation for Val.lazy() is provided. It calls the
given Supplier only once without notifing any observers. It is null-safe, but
not thread-safe.

An API is suggested for Val.lazyAsync(). Three variants are given. One merely
gives a Callable to be executed in some other thread. Another gives the
Executor to be used. Finally, a Supplier for a CompletionStage can be used as
well. Together, these should provide for every case from maximum usability to
maximum control.