Adds delayed async Val
A delayed version of the async Val is added. If, while waiting, all
observers are removed, the calculation will not be triggered. Once it has
been started, the unsubscription will be ignored.

The ScheduledExecutorServiceTimer gets moved into util and is set to
public. This was done because a) it is useful for this feature, but was
not visible, and b) it provides a nice abstraction and a useful
implementation of Timer besides FxTimer.