Renames forgetful to retainingLast
The method forgetful has a misleading name. For one, the defining property
is not so much that events are forgotten, but that the latest event will
be remembered. Also it is the only pair of *** and ***When methods for
suppressible streams where the names do not match. This should be
corrected.